backend/requirements.txt
alembic>=1.12.0
annotated-types>=0.5.0
anyio>=3.7.0
bcrypt>=4.0.0
blinker>=1.6.0
certifi>=2023.0.0
charset-normalizer>=3.0.0
click>=8.0.0
distro>=1.8.0
Flask>=3.0.0
Flask-Bcrypt>=1.0.0
flask-cors>=4.0.0
Flask-JWT-Extended>=4.5.0
Flask-Migrate>=4.0.0
Flask-SQLAlchemy>=3.0.0
h11>=0.14.0
httpcore>=0.17.0
httpx>=0.24.0
idna>=3.4.0
itsdangerous>=2.1.0
Jinja2>=3.1.0
jiter>=0.8.0
Mako>=1.2.0
MarkupSafe>=2.1.0
openai>=1.0.0
pydantic>=2.0.0
pydantic_core>=2.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
requests>=2.28.0
sniffio>=1.3.0
SQLAlchemy>=2.0.0
tqdm>=4.64.0
typing_extensions>=4.5.0
urllib3>=1.26.0
Werkzeug>=3.0.0
# AI dependencies (simplified)
langchain>=0.1.0
langchain-openai>=0.0.5
# Anthropic MCP dependencies
anthropic>=0.7.0